requirements.txt
python-dotenv
aiogram
psycopg2-binary